maybe last attempt to fix codacy

+++ projects/cm_mcu/LocalTasks.c
@@ -676,24 +676,24 @@ uint8_t getFFpresentbit(const uint8_t i)
   configASSERT(i < NFIREFLIES);
   if (i < NFIREFLIES_IT_F1) {
     j = i;
-    sel_bit_mask = (uint64_t)1 << (uint32_t)j;
+    sel_bit_mask = (uint8_t)1 << (uint8_t)j;
     val = (ffl12_f1_args.present_bit_mask & sel_bit_mask) && 1;
   }
 
   else if (NFIREFLIES_IT_F1 <= i && i < NFIREFLIES_IT_F1 + NFIREFLIES_DAQ_F1) {
     j = i - NFIREFLIES_IT_F1;
-    sel_bit_mask = (uint64_t)1 << (uint32_t)j;
+    sel_bit_mask = (uint8_t)1 << (uint8_t)j;
     val = (ffldaq_f1_args.present_bit_mask & sel_bit_mask) && 1;
   }
 
   else if (NFIREFLIES_F1 <= i && i < NFIREFLIES_F1 + NFIREFLIES_IT_F2) {
     j = i - NFIREFLIES_F1;
-    sel_bit_mask = (uint64_t)1 << (uint32_t)j;
+    sel_bit_mask = (uint8_t)1 << (uint8_t)j;
     val = (ffl12_f2_args.present_bit_mask & sel_bit_mask) && 1;
   }
   else {
     j = i - (NFIREFLIES_F1 + NFIREFLIES_IT_F2);
-    sel_bit_mask = (uint64_t)1 << (uint32_t)j;
+    sel_bit_mask = (uint8_t)1 << (uint8_t)j;
     val = (ffldaq_f2_args.present_bit_mask & sel_bit_mask) && 1;
   }
